add skeleton for envelope error handling.
