+++ aea/dialogue/base.py
@@ -27,10 +27,9 @@ This module contains the classes required for dialogue management.
 """
 
 from abc import abstractmethod
-from typing import Dict, List, TYPE_CHECKING
+from typing import Dict, List
 
-if TYPE_CHECKING:
-    from aea.protocols.base.message import Message
+from aea.protocols.base.message import Message
 
 
 class DialogueLabel:
@@ -90,9 +89,9 @@ class Dialogue:
         """
         self._dialogue_label = dialogue_label
         self._is_self_initiated = dialogue_label.dialogue_opponent_pbk is not dialogue_label.dialogue_starter_pbk
-        self._outgoing_messages = []  # type: List['Message']
-        self._outgoing_messages_controller = []  # type: List['Message']
-        self._incoming_messages = []  # type: List['Message']
+        self._outgoing_messages = []  # type: List[Message]
+        self._outgoing_messages_controller = []  # type: List[Message]
+        self._incoming_messages = []  # type: List[Message]
 
     @property
     def dialogue_label(self) -> DialogueLabel:
@@ -142,7 +141,7 @@ class Dialogues:
         return self._dialogues
 
     @abstractmethod
-    def is_permitted_for_new_dialogue(self, msg: 'Message', known_pbks: List[str]) -> bool:
+    def is_permitted_for_new_dialogue(self, msg: Message, known_pbks: List[str]) -> bool:
         """
         Check whether an agent message is permitted for a new dialogue.
 
@@ -153,7 +152,7 @@ class Dialogues:
         """
 
     @abstractmethod
-    def is_belonging_to_registered_dialogue(self, msg: 'Message', agent_pbk: str) -> bool:
+    def is_belonging_to_registered_dialogue(self, msg: Message, agent_pbk: str) -> bool:
         """
         Check whether an agent message is part of a registered dialogue.
 
@@ -164,7 +163,7 @@ class Dialogues:
         """
 
     @abstractmethod
-    def get_dialogue(self, msg: 'Message', agent_pbk: str) -> Dialogue:
+    def get_dialogue(self, msg: Message, agent_pbk: str) -> Dialogue:
         """
         Retrieve dialogue.
