+++ aea/protocols/default/message.py
@@ -42,6 +42,8 @@ class DefaultMessage(Message):
 
     class ErrorCode(Enum):
         UNSUPPORTED_PROTOCOL = -10001
+        DECODING_ERROR = -10002
+        INVALID_MESSAGE = -10003
 
     def __init__(self, type: Optional[Type] = None,
                  **kwargs):
